Add diamond economy as a possible economy type

+++ src/main/java/unsafedodo/guishop/economy/IEconomyService.java
@@ -4,9 +4,9 @@ import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 
 public interface IEconomyService {
-    boolean add(UUID uuid, double amount);
-    boolean remove(UUID uuid, double amount);
-    boolean hasEnoughMoney(UUID uuid, double amount) throws ExecutionException, InterruptedException;
-    double getBalance(UUID uuid) throws ExecutionException, InterruptedException;
-    boolean transfer(UUID sender, UUID receiver, double amount);
+    boolean add(UUID uuid, long amount);
+    boolean remove(UUID uuid, long amount);
+    boolean hasEnoughMoney(UUID uuid, long amount) throws ExecutionException, InterruptedException;
+    long getBalance(UUID uuid) throws ExecutionException, InterruptedException;
+    boolean transfer(UUID sender, UUID receiver, long amount);
 }
